# serveur-REST-depenses-v0.2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Serveur des dépenses partagées",
        "description": "Utilisation d'un serveur REST",
        "contact": {
            "email": "fred.hemery@univ-artois.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Le serveur pour les dépenses partagées"
        }
    ],
    "paths": {
...
        "/membres/{id}": {
            "get": {
                "tags": [
                    "Membres"
                ],
                "summary": "Renvoie les informations d'un membre",
                "description": "Renvoie les informations d'un membre",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du membre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "membre response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "required": [
                                                "id"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int64"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "nom": {
                                                    "type": "string"
                                                },
                                                "prenom": {
                                                    "type": "string"
                                                },
                                                "avatar": {
                                                    "type": "string"
                                                },
                                                "password": {
                                                    "type": "string"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "achats": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "montant": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "date_achat": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "depenses": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "acheteur_id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "acheteur": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "quote_part": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "date_depense": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:membres",
                            "read:membres"
                        ]
                    }
                ]
            }
        },
        "/membres": {
            "get": {
                "tags": [
                    "Membres"
                ],
                "summary": "Get list of membres",
                "description": "Renvoie la liste des membres",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            },
                                                            "nom": {
                                                                "type": "string"
                                                            },
                                                            "prenom": {
                                                                "type": "string"
                                                            },
                                                            "avatar": {
                                                                "type": "string"
                                                            },
                                                            "password": {
                                                                "type": "string"
                                                            },
                                                            "solde": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "achats": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer",
                                                                            "format": "int64"
                                                                        },
                                                                        "description": {
                                                                            "type": "string"
                                                                        },
                                                                        "montant": {
                                                                            "type": "number",
                                                                            "format": "float"
                                                                        },
                                                                        "date_achat": {
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "depenses": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer",
                                                                            "format": "int64"
                                                                        },
                                                                        "acheteur_id": {
                                                                            "type": "integer",
                                                                            "format": "int64"
                                                                        },
                                                                        "acheteur": {
                                                                            "type": "string"
                                                                        },
                                                                        "description": {
                                                                            "type": "string"
                                                                        },
                                                                        "quote_part": {
                                                                            "type": "number",
                                                                            "format": "float"
                                                                        },
                                                                        "date_depense": {
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
...
}